scaler/requirements.txt
docker
pyyaml
flask
numpy
matplotlib
